\name{TSPC-package}
\alias{TSPC-package}
\alias{TSPC}
\docType{package}
\title{
TSPC
}
\description{
Performs survival and quantitative outcome using time-course gene expression, described in the following paper:Zhang Y, Tibshirani RJ, Davis RW. Predicting patient survival from longitudinal gene expression. Stat Appl Genet Mol Biol. 2010;9(1):Article41. Epub 2010 Nov 22.
}
\details{
\tabular{ll}{
Package: \tab TSPC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-25\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yuping Zhang <yupingz@stanford.edu>
}
\references{
Zhang Y, Tibshirani RJ, Davis RW. Predicting patient survival from longitudinal gene expression. Stat Appl Genet Mol Biol. 2010;9(1):Article41. Epub 2010 Nov 22.
}
\keyword{  regression }
\keyword{ survival }
