% --- Source file: man/tspc.internal.Rd ---
\name{superpc-internal}
\title{Internal superpc functions}
\alias{balanced.folds}  
\alias{cor.func}  
\alias{cox.coef.func}  
\alias{coxfunc}  
\alias{coxscor} 
\alias{coxstuff}  
\alias{coxvar}  
\alias{error.bars}  
\alias{jitter}  
\alias{lm.coef.func}
\alias{max.na} 
\alias{mean.na}  
\alias{min.na}  
\alias{multiply.func}  
\alias{mysvd}  
\alias{ocoxvar}  
\alias{permute.rows} 
\alias{se.na} 

\description{Internal TSPC functions}
\author{Yuping Zhang}
\details{
  These are not to be called by the user.
}
\keyword{internal}

